polaris_integrations/requirements.txt
# Polaris Integrations requirements.txt (para Python 3.10.x)

TTS==0.22.0
faster-whisper==0.10.0
transformers==4.37.2
tokenizers==0.15.1
huggingface-hub==0.23.0

numpy==1.22.0

python-telegram-bot==20.8
fastapi==0.110.1
uvicorn[standard]==0.29.0
httpx==0.26.0

python-dotenv==1.0.1
requests==2.31.0
prometheus-client==0.20.0
moviepy
groq
